requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
openai>=1.0.0
python-dotenv==1.0.1
pydantic==2.9.2
sqlalchemy==2.0.35
itsdangerous==2.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.22
